improve release

+++ rel/rpm/centos/SPECS/emsbus.spec
@@ -1,7 +1,7 @@
 Summary: A service-oriented bus developed in Erlang/OTP by Everton de Vargas Agilar
 Name: ems-bus
 Version: 1.0.12
-Release: el7.centos
+Release: centos.7
 License: GPL
 Group: System/Utilities
 URL: https://github.com/erlangms
@@ -10,6 +10,9 @@ Packager: ErlangMS Team <evertonagilar@unb.br>
 BuildRoot: %{_tmppath}/%{name}-{%version}
 Source0: %{name}-%version.tar.gz
 Provides: ems-bus
+AutoReqProv: no
+Requires: /bin/bash /bin/sh /bin/sh /bin/sh /usr/bin/env bash >= 2.0 libc.so.6()(64bit)  libc.so.6(GLIBC_2.10)(64bit) libc.so.6(GLIBC_2.11)(64bit) libc.so.6(GLIBC_2.14)(64bit) libc.so.6(GLIBC_2.15)(64bit) libc.so.6(GLIBC_2.2.5)(64bit) libc.so.6(GLIBC_2.3)(64bit) libc.so.6(GLIBC_2.3.2)(64bit) libc.so.6(GLIBC_2.3.4)(64bit) libc.so.6(GLIBC_2.4)(64bit) libc.so.6(GLIBC_2.7)(64bit) libc.so.6(GLIBC_2.8)(64bit) libcrypto.so.10()(64bit) libcrypto.so.10(OPENSSL_1.0.1)(64bit) libcrypto.so.10(OPENSSL_1.0.1_EC)(64bit) libcrypto.so.10(libcrypto.so.10)(64bit) libdl.so.2()(64bit) libdl.so.2(GLIBC_2.2.5)(64bit) libgcc_s.so.1()(64bit) libgcc_s.so.1(GCC_3.0)(64bit) libm.so.6()(64bit) libm.so.6(GLIBC_2.2.5)(64bit) libodbc.so.2()(64bit) libpthread.so.0()(64bit) libpthread.so.0(GLIBC_2.12)(64bit) libpthread.so.0(GLIBC_2.2.5)(64bit) libpthread.so.0(GLIBC_2.3.2)(64bit) librt.so.1()(64bit) librt.so.1(GLIBC_2.2.5)(64bit) libstdc++.so.6()(64bit) libstdc++.so.6(CXXABI_1.3)(64bit) libtinfo.so.5()(64bit) libutil.so.1()(64bit) libutil.so.1(GLIBC_2.2.5)(64bit) libz.so.1()(64bit) libz.so.1(ZLIB_1.2.2)(64bit) rpmlib(CompressedFileNames) <= 3.0.4-1 rpmlib(FileDigests) <= 4.6.0-1 rpmlib(PartialHardlinkSets) <= 4.0.4-1 rpmlib(PayloadFilesHavePrefix) <= 4.0-1 rtld(GNU_HASH) rpmlib(PayloadIsXz) <= 5.2-1
+
 
 %description
 ErlangMS is a enterprise service bus developed in Erlang/OTP to facilitate the integration of systems through a service-oriented approach for the systems of the University of Brazilia. 
@@ -26,6 +29,8 @@ by graduate student Everton Vargas Agilar.
 
 %install
 
+  sudo systemctl daemon-reload
+  sudo systemctl stop ems-bus.service  > /dev/null 2>&1 || true
   mkdir -p $RPM_BUILD_ROOT
   cp -R $RPM_SOURCE_DIR/* $RPM_BUILD_ROOT/
 
@@ -36,6 +41,7 @@ by graduate student Everton Vargas Agilar.
   USER_EMS_BUS=erlangms
   GROUP_EMS_BUS=erlangms
   HOME_EMS_BUS=/var/opt/erlangms
+  USER_CREATED="false"
 
   # database and log path
   mkdir -p /usr/lib/ems-bus/priv/db > /dev/null 2>&1
@@ -53,6 +59,7 @@ by graduate student Everton Vargas Agilar.
 								--password $password_crypt \
 								--comment "User do barramento ERLANGMS" $USER_EMS_BUS  > /dev/null 2>&1
       echo "User erlangms created with admin passwd. Change after installation!"
+      USER_CREATED="true"
   fi
   
   # The starters need to be Suid root.
@@ -75,10 +82,13 @@ by graduate student Everton Vargas Agilar.
   fi
   ln -s /usr/lib/ems-bus/ $HOME_EMS_BUS/ems-bus > /dev/null 2>&1
   
-  # remove o .odbc.ini original criado pelo useradd e cria um link para o priv/conf/odbc.ini
-  rm -f $HOME_EMS_BUS/.odbc.ini > /dev/null 2>&1
-  ln -s /usr/lib/ems-bus/priv/conf/odbc.ini $HOME_EMS_BUS/.odbc.ini > /dev/null 2>&1
 
+  # It only changes the $ HOME_EMS_BUS/.odbc.ini file when the user was created
+  if [ "$USER_CREATED" == "true" ]; then
+	  rm -f $HOME_EMS_BUS/.odbc.ini > /dev/null 2>&1
+	  cp /usr/lib/ems-bus/priv/conf/odbc.ini $HOME_EMS_BUS/.odbc.ini > /dev/null 2>&1
+  fi
+  
   chown -Rf $USER_EMS_BUS:$GROUP_EMS_BUS /var/opt/erlangms > /dev/null 2>&1	
   chown -Rf $USER_EMS_BUS:$GROUP_EMS_BUS /var/opt/erlangms/.erlangms > /dev/null 2>&1	
 
@@ -115,6 +125,7 @@ by graduate student Everton Vargas Agilar.
   chown root:root /etc/sudoers.d/ems-bus.sudoers > /dev/null 2>&1
   chown root:root /etc/systemd/system/ems-bus.service > /dev/null 2>&1
 
+
   # config /etc/odbcinst.ini if necessary for FreeTDS SQL-server driver
   JTDS_ENTRY_CONF=$(sed -rn '/\[FreeTDS\]/, /(^$|^#)/p' /etc/odbcinst.ini 2> /dev/null)
   if [ -z "$JTDS_ENTRY_CONF" ]; then
@@ -127,6 +138,7 @@ by graduate student Everton Vargas Agilar.
 	echo " " >> /etc/odbcinst.ini 
   fi
 
+
   # ldconfig
   /sbin/ldconfig  > /dev/null 2>&1 || true
 
@@ -136,19 +148,22 @@ by graduate student Everton Vargas Agilar.
   systemctl enable ems-bus.service  > /dev/null 2>&1 || true
   systemctl daemon-reload  > /dev/null 2>&1 || true
   systemctl start ems-bus.service  > /dev/null 2>&1 || true
-
+  systemctl daemon-reload > /dev/null 2>&1
 
 %postun 
 
 	# pare e desative o serviço systemctl
 	systemctl stop ems-bus.service  > /dev/null 2>&1 || true
 	systemctl disable ems-bus.service > /dev/null 2>&1 || true
+	systemctl daemon-reload
 
 	# remove user
 	#groupdel erlangms > /dev/null 2>&1 || true
 	#userdel erlangms > /dev/null 2>&1 || true
 
 	/sbin/ldconfig
+	
+
